requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.10.2; python_version >= '3.7'
attrs==22.2.0; python_version >= '3.6'
black==23.3.0
cfgv==3.3.1; python_full_version >= '3.6.1'
click==8.1.3; python_version >= '3.7'
distlib==0.3.6
filelock==3.10.7; python_version >= '3.7'
flake8==6.0.0
flasgger==0.9.5
flask-cors==3.0.10
flask-marshmallow==0.14.0
flask-migrate==4.0.4
flask-sqlalchemy==3.0.3
flask==2.2.3
greenlet==2.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
identify==2.5.22; python_version >= '3.7'
importlib-metadata==6.1.0; python_version < '3.10'
isort==5.12.0
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jsonschema==4.17.3; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
marshmallow==3.19.0
mccabe==0.7.0; python_version >= '3.6'
mistune==2.0.5
mypy-extensions==1.0.0; python_version >= '3.5'
mysql-connector-python==8.0.29
nodeenv==1.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.0; python_version >= '3.7'
pathspec==0.11.1; python_version >= '3.7'
platformdirs==3.2.0; python_version >= '3.7'
pre-commit==3.2.1
protobuf==4.22.1; python_version >= '3.7'
pycodestyle==2.10.0; python_version >= '3.6'
pyflakes==3.0.1; python_version >= '3.6'
pyjwt==2.6.0
pyrsistent==0.19.3; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.7; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.5.0; python_version < '3.10'
virtualenv==20.21.0; python_version >= '3.7'
werkzeug==2.2.3; python_version >= '3.7'
zipp==3.15.0; python_version >= '3.7'
